Moved View to its own component file
