+++ mobile/app/[...missing].tsx
@@ -1,7 +1,8 @@
 import { Link, Stack } from 'expo-router'
 import { StyleSheet } from 'react-native'
 
-import { Text, View } from '../components/Themed'
+import { Text } from '../components/Themed'
+import { View } from "../components/View";
 
 export default function NotFoundScreen() {
 	return (
