+++ mobile/components/Themed.tsx
@@ -3,16 +3,15 @@
  * https://docs.expo.io/guides/color-schemes/
  */
 
-import { Text as DefaultText, useColorScheme, View as DefaultView } from 'react-native'
+import { Text as DefaultText, useColorScheme } from 'react-native'
 import Colors from '../constants/Colors'
 
-type ThemeProps = {
+export type ThemeProps = {
 	lightColor?: string
 	darkColor?: string
 }
 
 export type TextProps = ThemeProps & DefaultText['props']
-export type ViewProps = ThemeProps & DefaultView['props']
 
 export function useThemeColor(
 	props: { light?: string, dark?: string },
@@ -35,9 +34,3 @@ export function Text(props: TextProps) {
 	return <DefaultText style={[{ color }, style]} {...otherProps} />
 }
 
-export function View(props: ViewProps) {
-	const { style, lightColor, darkColor, ...otherProps } = props
-	const backgroundColor = useThemeColor({ light: lightColor, dark: darkColor }, 'background')
-
-	return <DefaultView style={[{ backgroundColor }, style]} {...otherProps} />
-}
